Split ExtensionService into two classes
