+++ src/NUnitEngine/nunit-agent/Program.cs
@@ -87,7 +87,7 @@ namespace NUnit.Agent
             };
 
             // Custom Service Initialization
-            engine.Services.Add(new ExtensionService(isRunningOnAgent: true));
+            engine.Services.Add(new ExtensionService());
             engine.Services.Add(new DriverService());
 
             // Initialize Services
